Feat(web-react): Storybook preview update #DS-331
