+++ packages/web-react/src/components/Modal/Modal.stories.tsx
@@ -14,4 +14,3 @@ export default {
 } as ComponentMeta<typeof Modal>;
 
 export { default as Modal } from './stories/Modal';
-export { default as ModalComposed } from './stories/ModalComposed';
